\name{PoTwo}
\alias{PoTwo}
\title{Square of a number}
\usage{
PoTwo(x)
}
\description{
Gives the square of the input integer.
}
\arguments{
  \item{x}{
    an integer
  }
}
\value{
  The square of the input integer is computed and returned as a numeric value. 
}
\author{
  Louis Preisig, \email{louis.preisig@uzh.ch}
}
\examples{
  x = 3
  PoTwo(x)
  
}
